pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ved Patel <vedpatel3395@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.120.4"
mongo = "^0.2.0"
uvicorn = "^0.38.0"
pydantic-settings = "^2.11.0"
openai = "^2.6.1"
pyyaml = "^6.0.3"
motor = "^3.7.1"
python-multipart = "^0.0.20"
python-docx = "^1.1.2"
docx = "^0.2.4"
langchain = "^1.0.5"
langchain-openai = "^1.0.2"
langchain-community = "^0.4.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
